Moved controllers / models to seperate folders
