+++ controllers/controllers.js
@@ -1,0 +1,4 @@
+const { getCategories  } = require('./category-controllers.js')
+const { getReviewById } = require('./review-controllers.js')
+
+module.exports = {getReviewById, getCategories  }
