+++ models/category-models.js
@@ -1,0 +1,8 @@
+const db = require('./../db/connection.js');
+exports.selectCategories = () => {
+  return db
+  .query('SELECT * FROM categories;')
+  .then((result)=>{
+    return result.rows
+  })
+}
